moved all consumer/producer functions to the PulsarController
