+++ src/main/scala/com/raphtory/core/components/partition/QueryExecutor.scala
@@ -40,15 +40,20 @@ class QueryExecutor(
   var receivedMessageCount: Int = 0
   var votedToHalt: Boolean      = false
 
-  private val taskManager: Producer[Array[Byte]]          = toQueryHandlerProducer(jobID)
-  private val neighbours: Map[Int, Producer[Array[Byte]]] = toQueryExecutorProducers(jobID)
+  private val taskManager: Producer[Array[Byte]] = pulsarController.toQueryHandlerProducer(jobID)
+
+  private val neighbours: Map[Int, Producer[Array[Byte]]] =
+    pulsarController.toQueryExecutorProducers(jobID)
   var cancelableConsumer: Option[Consumer[Array[Byte]]]   = None
 
   override def run(): Unit = {
     logger.debug(s"Job '$jobID' at Partition '$partitionID': Starting query executor consumer.")
 
     taskManager sendAsync serialise(ExecutorEstablished(partitionID))
-    cancelableConsumer = Some(startQueryExecutorConsumer(Schema.BYTES, partitionID, jobID))
+    cancelableConsumer = Some(
+            pulsarController
+              .startQueryExecutorConsumer(Schema.BYTES, partitionID, jobID, messageListener())
+    )
   }
 
   override def stop(): Unit = {
